alarino-backend/data/create_tables.py
#create_tables.py

from db_models import db
from flask import Flask
from dotenv import load_dotenv
import os

load_dotenv()

app = Flask(__name__)
app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("DATABASE_URL")
app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
db.init_app(app)

with app.app_context():
    print("trying")
    db.create_all()  # this creates all tables based on the models
    print("✅ Tables created successfully.")
